requirements.txt
beautifulsoup4>=4.12.3,<5.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.1.0,<3.0.0
psutil>=5.9.0,<6.0.0
pyyaml>=6.0,<7.0
requests>=2.31.0,<3.0.0
schedule>=1.2.0,<2.0.0
selenium>=4.15.0,<5.0.0
seleniumbase>=4.25.0,<5.0.0
sqlalchemy>=2.0.0,<3.0.0
tabulate>=0.9.0,<0.10.0
